requirements.txt
Django
djangorestframework
python-dotenv
drf-yasg
psycopg2